requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.3
django-phone-field==1.8.1
django-tinymce==3.5.0
django-tinymce4==3.0.2
django-versatileimagefield==2.2
Pillow==9.3.0
python-decouple==3.6
python-magic==0.4.27
sqlparse==0.4.3
